+++ kotlin-sqlobject/src/main/kotlin/org/jdbi/v3/sqlobject/kotlin/KotlinSqlObjectPlugin.kt
@@ -17,10 +17,11 @@ import org.jdbi.v3.core.Jdbi
 import org.jdbi.v3.core.spi.JdbiPlugin
 import org.jdbi.v3.sqlobject.Handlers
 import org.jdbi.v3.sqlobject.SqlObjects
+import org.jdbi.v3.sqlobject.SqlObjectPlugin
 
-class KotlinSqlObjectPlugin : JdbiPlugin {
-
+class KotlinSqlObjectPlugin : JdbiPlugin.Singleton() {
     override fun customizeJdbi(jdbi: Jdbi) {
+        jdbi.installPlugin(SqlObjectPlugin());
         jdbi.configure(SqlObjects::class.java, { c -> c.defaultParameterCustomizerFactory = KotlinSqlStatementCustomizerFactory() })
         jdbi.configure(Handlers::class.java, { c -> c.register(KotlinDefaultMethodHandlerFactory()) })
     }